Rescue from StandardError in #find_articles
Instead of listing all the possible HTTP errors we can get (which are a
lot if we include Net::HTTP errors, RestClient errors and others related
to timeouts and what not) we're rescuing from a possible StandardError.

This is not as bad as it seems since we're running a limited amount of
retries and if the error persists we finally raise it.